requirements.txt
fastapi
uvicorn
python-multipart
easyocr
numpy
Pillow
opencv-python-headless